nlp/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
bioc==2.0.post4
blis==0.7.9
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.0.4
cymem==2.0.11
emoji==2.12.0
filelock==3.12.2
Flask==2.2.5
fsspec==2023.1.0
gast==0.3.3
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==2.10.0
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==6.7.0
intervaltree==3.1.0
itsdangerous==2.1.2
Jinja2==3.1.5
joblib==1.3.2
jsonlines==3.1.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
lxml==5.3.1
Markdown==3.4.4
MarkupSafe==2.1.5
murmurhash==1.0.12
numpy==1.18.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pathy==0.10.3
preshed==3.0.9
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.8.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sacremoses==0.0.53
scipy==1.4.1
six==1.17.0
smart-open==6.4.0
sortedcontainers==2.4.0
spacy==3.2.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stanza==1.4.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-addons==0.12.1
tensorflow-estimator==2.3.0
termcolor==2.3.0
thinc==8.0.17
tokenizers==0.12.1
torch==1.13.1
tqdm==4.67.1
transformers==4.18.0
typeguard==2.13.3
typer==0.4.2
typing_extensions==4.7.1
urllib3==2.0.7
wasabi==0.10.1
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
